Make camera position smoothing apply to 'Body' Node3D
Less lines of code for (almost) the same result.
Allows for more than just a camera (a player mesh, for example) to follow the interpolated path while climbing stairs
Avoids the use of 'set_as_top_level()' that leads to confusing script behavior.

The only real bug I could find is on top of the catwalk if you run into the top step's corners and the railing. The camera (and presumably the entire 'Body' Node) oscillates up and down despite no collider movement. It has something to do with when 'step_result.diff_position' is subtracted in my version, but haven't figured out a fix yet.